Refactor from sim-utils into macro-utils, make chem-utils stub
