+++ src/utils/sim-utils.ts
@@ -619,18 +619,3 @@ export const draggableLeafTypes = [
   LeafType.oak0, LeafType.oak1, LeafType.oak2, LeafType.oak3, LeafType.oak4, LeafType.oak5, LeafType.oak6,
   LeafType.maple0, LeafType.maple1, LeafType.maple2, LeafType.maple3, LeafType.maple4, LeafType.maple5, LeafType.maple6
 ];
-
-// PTI
-export interface PTIRatingLevel {
-  color: string;
-  label: string;
-  range: string;
-  min: number;
-}
-
-export const PTIRatingLevels: PTIRatingLevel[] = [
-  {color: "#a4f9be", label: t("PTI.RATING.LEVEL0.LABEL"), range: t("PTI.RATING.LEVEL0.RANGE"), min: 24},
-  {color: "#94e5ff", label: t("PTI.RATING.LEVEL1.LABEL"), range: t("PTI.RATING.LEVEL1.RANGE"), min: 17},
-  {color: "#cccccc", label: t("PTI.RATING.LEVEL2.LABEL"), range: t("PTI.RATING.LEVEL2.RANGE"), min: 11},
-  {color: "#ffacac", label: t("PTI.RATING.LEVEL3.LABEL"), range: t("PTI.RATING.LEVEL3.RANGE"), min: 0}
-];
